+++ crates/wdl-engine/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Implement the string array functions from the WDL standard library ([#255](https://github.com/stjude-rust-labs/wdl/pull/255)).
 * Replaced the `Value::from_json` method with `Value::deserialize` which allows
   for deserialization from any self-describing data format; a method for
   serializing a value was also added ([#254](https://github.com/stjude-rust-labs/wdl/pull/254)).